Develop d

+++ src/main/java/com/efitops/basesetup/service/AmountInWordsConverterService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 @Service
 public interface AmountInWordsConverterService {
 
-	String convert(Long amount);
+	String convert(BigDecimal amount);
